Remove previous ANN implementation, refactor
